chore: moves cmd logic to pkg/
