+++ e2e/infra/istio_setup.go
@@ -4,8 +4,8 @@ import (
 	"os"
 	"time"
 
-	"github.com/maistra/istio-workspace/cmd/ike/cmd"
-	"github.com/maistra/istio-workspace/cmd/ike/config"
+	"github.com/maistra/istio-workspace/pkg/cmd"
+	"github.com/maistra/istio-workspace/pkg/cmd/config"
 
 	"github.com/onsi/gomega"
 )
