+++ pkg/cmd/develop.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/maistra/istio-workspace/cmd/ike/config"
+	"github.com/maistra/istio-workspace/pkg/cmd/config"
 
 	gocmd "github.com/go-cmd/cmd"
 	"github.com/spf13/cobra"
