+++ pkg/cmd/internal/session/session_suite_test.go
