+++ pkg/cmd/internal/session/session_test.go
@@ -1,7 +1,7 @@
 package session_test
 
 import (
-	"github.com/maistra/istio-workspace/cmd/ike/session"
+	"github.com/maistra/istio-workspace/pkg/cmd/internal/session"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
