+++ pkg/cmd/internal/watch/watch_test.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"path/filepath"
 
-	"github.com/maistra/istio-workspace/cmd/ike/watch"
+	"github.com/maistra/istio-workspace/pkg/cmd/internal/watch"
 
 	"github.com/fsnotify/fsnotify"
 	. "github.com/onsi/ginkgo"
