+++ pkg/cmd/session_func.go
@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	"github.com/maistra/istio-workspace/cmd/ike/session"
+	"github.com/maistra/istio-workspace/pkg/cmd/internal/session"
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
