+++ pkg/cmd/session_func_test.go
@@ -1,7 +1,7 @@
 package cmd_test
 
 import (
-	. "github.com/maistra/istio-workspace/cmd/ike/cmd"
+	. "github.com/maistra/istio-workspace/pkg/cmd"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
